Had time to work on CSS styles
Finally had a chance to work on this project again. Worked on the CSS and got most of the left side of the desktop layout done. Need to still work on country and postal code sections. Also still need to work on the shopping cart. Will use JS for the first time on one of these projects:)